docs: reworked documentation, removed commented code, slight refactors.
